+++ frontend/src/App.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { BrowserRouter, Route, Routes, Navigate } from 'react-router-dom';
 
 import './App.css';
+import CourseCreate from './containers/CourseCreate/CourseCreate';
 import CourseList from './containers/CourseList/CourseList';
-import CourseSearch from './containers/CourseSearch/CourseSearch';
 import MainPage from './containers/MainPage/MainPage';
 import CourseDetail from './DetailPage';
 import Register from './registerPage';
@@ -15,7 +15,7 @@ function App() {
         <Routes>
           <Route path="/courses" element={<CourseList />} />
           <Route path="/course/:id" element={<CourseDetail />} />
-          <Route path="search" element={<CourseSearch />} />
+          <Route path="search" element={<CourseCreate />} />
           <Route path="/register/" element={<Register />} />
           <Route path="/" element={<MainPage />} />
           <Route path="*" element={<Navigate replace to="/" />} />
