+++ frontend/src/containers/CourseCreate/CourseCreate.module.scss
@@ -1,5 +1,5 @@
 .Input {
-  height: 30px;
+  height: 20px;
 }
 
 .Input:focus {
